requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
scikit-learn
streamlit
transformers
torch
python-dotenv
requests
